requirements.txt
beautifulsoup4==4.7.1
better-profanity==0.1
certifi==2018.11.29
chardet==3.0.4
Click==7.0
dnspython==1.16.0
eventlet==0.24.1
Flask==1.0.2
Flask-SocketIO==3.1.1
greenlet==0.4.15
idna==2.8
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10
markovify==0.7.1
MarkupSafe==1.1.0
monotonic==1.5
oauthlib==3.0.0
PySocks==1.6.8
python-dotenv==0.10.1
python-engineio==3.2.3
python-socketio==3.1.1
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
soupsieve==1.7.2
tweepy==3.7.0
Unidecode==1.0.23
urllib3==1.24.1
Werkzeug==0.14.1
